chore: move data into data/ folder
Clean up the root directory somewhat
