+++ util/manual_utils.ts
@@ -1,11 +1,13 @@
 // Copyright 2022 the Deno authors. All rights reserved. MIT license.
 
+import compareVersions from "$tiny-version-compare";
+import { getSourceURL } from "./registry_utils.ts";
+import VERSIONS from "@/data/versions.json" assert { type: "json" };
+
 const githubBasepath = "https://raw.githubusercontent.com/denoland/manual/";
 const oldDocpath = "https://github.com/denoland/deno/blob/";
 const docpath = "https://github.com/denoland/manual/blob/";
-import VERSIONS from "../versions.json" assert { type: "json" };
-import { getSourceURL } from "./registry_utils.ts";
-import compareVersions from "$tiny-version-compare";
+
 export const versions = VERSIONS.cli;
 
 export interface TableOfContents {
